+++ tests/golang/simple.go
@@ -1,0 +1,34 @@
+package main
+
+import (
+	"./cycletls"
+	"log"
+	"runtime"
+	"time"
+	// "net/http"
+)
+
+func main() {
+	runtime.GOMAXPROCS(runtime.NumCPU())
+
+	start := time.Now()
+	defer func() {
+		log.Println("Execution Time: ", time.Since(start))
+	}()
+	client := cycletls.Init()
+	resp, err := client.Do("https://httpbin.org/deflate", cycletls.Options{
+		Body:      "",
+		Ja3:       "771,4865-4867-4866-49195-49199-52393-52392-49196-49200-49162-49161-49171-49172-51-57-47-53-10,0-23-65281-10-11-35-16-5-51-43-13-45-28-21,29-23-24-25-256-257,0",
+		UserAgent: "Mozilla/5.0 (X11; Ubuntu; Linux x86_64; rv:87.0) Gecko/20100101 Firefox/87.0",
+				Timeout:   30,
+
+	}, "GET")
+	if err != nil {
+		log.Print("Request Failed: " + err.Error())
+	}
+	log.Println(resp.Body)
+	if resp.JSONBody()["deflated"] != true {
+		log.Println("Deflated not found in Reponse Body: %s", resp.Body)
+	}
+
+}